Resolved comments from Mr. Dias

+++ 2023-Robot/src/main/test/java/frc/team670/robot/commands/arm/InverseKinematicsTest.java
@@ -7,7 +7,7 @@ public class InverseKinematicsTest {
         for(int i = 0; i <= 9; i++){
         
             //ArmState.getVal(i);    
-            System.out.println(ArmState.getVal(i)+" Shoulder Angle: "+ ArmState.getVal(i).getShoulderAngle()+" Elbow Angle: "+ArmState.getVal(i).getElbowAngle());        
+            System.out.println(ArmState.getVal(i)+" Shoulder Angle: "+ (int) ArmState.getVal(i).getShoulderAngle()+" Elbow Angle: "+ (int) ArmState.getVal(i).getElbowAngle());        
         }
         
     }
